refactor(utils): move unused exted_dict to test utils
